backend/requirements.txt
flask
flask-cors
pypdf
sentence-transformers
faiss-cpu
python-dotenv
requests
ollama
gunicorn
tqdm
numpy
